修改卡池ID类型为long

+++ Actions/CardReference.cs
@@ -4,7 +4,7 @@ namespace TouhouCardEngine
     [Serializable]
     public class CardReference
     {
-        public int cardPoolId;
+        public long cardPoolId;
         public int cardId;
     }
 }
